requirements.txt
anthropic>=0.40.0
groq>=0.9.0
python-dotenv>=1.0.0
nltk>=3.8.1
